% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{RegionExperience}
\alias{RegionExperience}
\title{Region data}
\format{A data frame with four variables:
   \code{Region}, \code{PolicyYear}
   , \code{NumPolicies}, \code{NumClaims}}
\usage{
RegionExperience
}
\description{
Consolidation of state data into various regions
}
\keyword{datasets}

