% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Utility.R
\name{CopySlides}
\alias{CopySlides}
\title{Copy Slides}
\usage{
CopySlides(SavePath, overwrite = TRUE, lib)
}
\arguments{
\item{SavePath}{Character string giving the name of the location where the slides should be copied.}

\item{overwrite}{Boolean indicating whether to overwrite any existing files. The default for this
parameter is TRUE.}
}
\description{
This function will copy slides to a directory specified by the user. Note that if the directory
already exists, the functino will issue a warning
}

